Use jsdom instead of phantomjs for linkifier test
